// web/src/api/model/badukBoardGameConfig.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * doljabi
 * doljabi project REST API를 정의한 문서입니다.
 * OpenAPI spec version: 1.0.0
 */

export interface BadukBoardGameConfig {
  /** @minimum 0 */
  fischer_time: number;
  /** @minimum 0 */
  main_time: number;
  /** @minimum 0 */
  overtime: number;
  /** @minimum 0 */
  remaining_overtime: number;
}